test: extract PathMode tests from write.rs into write_path_mode.rs
Closes #2401
